+++ src/app/shared/module/poe/price/item-price-prediction.ts
@@ -1,5 +1,6 @@
 export interface ItemPricePredictionResult {
     id: ItemPricePredictionResultId;
+    url: string;
     currency: string;
     min: number;
     max: number;